// jobup-frontend/src/app/generated-sources/openapi/model/authResponseDto.ts
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * DTO returned after successful authentication
 */
export interface AuthResponseDto { 
    token?: string;
    roles?: Array<string>;
    userId?: string;
    username?: string;
    email?: string;
}