// jobup-frontend/src/app/generated-sources/openapi/model/registerRequestDto.ts
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * DTO used to register a new user
 */
export interface RegisterRequestDto { 
    username: string;
    email: string;
    password: string;
}